+++ src/main/java/com/protrack/protrack_be/repository/FunctionRepository.java
@@ -3,4 +3,4 @@ package com.protrack.protrack_be.repository;
 import com.protrack.protrack_be.model.Function;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FunctionRepository extends JpaRepository<Function, Integer> {}
+public interface FunctionRepository extends JpaRepository<Function, Long> {}
