+++ src/main/java/com/protrack/protrack_be/repository/InvitationRepository.java
@@ -3,4 +3,4 @@ package com.protrack.protrack_be.repository;
 import com.protrack.protrack_be.model.Invitation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvitationRepository extends JpaRepository<Invitation, Integer> {}
+public interface InvitationRepository extends JpaRepository<Invitation, Long> {}
