+++ src/main/java/com/protrack/protrack_be/repository/ProjectRepository.java
@@ -3,4 +3,4 @@ package com.protrack.protrack_be.repository;
 import com.protrack.protrack_be.model.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository extends JpaRepository<Project, Integer> {}
+public interface ProjectRepository extends JpaRepository<Project, Long> {}
